Create README.md
current readme with instructions for installing required libraries and how to use the program from terminal line. Currently creates an html file on local device and opens this html in browser on system. Plan to later just make this a web app for the user to use in browser so that they do not need to create files on their local system.